Change structure of mobile navbar so it takes space on page
Previously, we needed to manually shift the contents of a page down because
the navbar did not take up space on the page. This changes its behavior so
it now does so and no longer needs manual adjustments to ensure this. Also
increased the logo size slightly and removed some unnecessary styles.