Fully implement sidebar according to figma design
I moved the arrow button that toggles the desktop sidebar to be outside of the sidebar div completely, this way it is much easier to position and isn't cut off by the boundaries of the div. I also added alternative color backgrounds for the hover effect of the sidebar links, but these colors are probably subject to change. Finally, I adjusted some padding so the sidebar looks right on all viewport sizes.